fix: input move in login, register components
